refactor chat route exports
